Bridge: Added function to enable bootloader after reset
Some microcrontrollers don't start the bootloader after a reset.
The new function is intended to let the microcontroller erase its
flash after checking a specific signal coming from the external
device without the need to press the erase button on the board.

The purpose is to enable a software update that does not require
a manual interaction with the board.